refactor: restructure CLI layer with unified CliCommand dispatch and rendering pipeline
All analyst-facing commands are now CliCommand variants. The previous split
between LibCommand and CliCommand dispatch loops is replaced by a single
command table and a single dispatch loop in the REPL.

Key changes:
- cli_commands.hpp, command_info.hpp, command_tables.hpp collapsed into
  commands.hpp and command_table.hpp
- All lib commands wrapped in corresponding Cli* structs (CliFilter,
  CliSort, CliDistinct, etc.)
- HeadCommand, TailCommand, PeekCommand, RowsCommand replaced by
  RecordsCommand with optional start/end/raw fields
- RowsResult renamed to RecordsResult
- RenderContext introduced to carry schema, terminal width, and output
  stream through the rendering pipeline
- RenderOptions introduced for per-call rendering preferences (DisplayMode)
- dispatch() helper unifies lib command execution and result rendering
- ExitRequested exception replaces bool return value for exit signalling
- Vertical and untruncated display modes implemented for peek, head, tail
- Command aliases supported in dispatch loop and help system

+++ hawk-cli/src/helpers/utils.cpp
@@ -70,8 +70,8 @@ std::optional<char> parse_delimiter(const std::string& str) {
     return std::nullopt;
 }
 
-std::uint8_t num_digits(std::uint64_t n) {
-    return (std::uint8_t)std::log10(n) + 1;
+std::size_t num_digits(std::uint64_t n) {
+    return (std::size_t)std::log10(n) + 1;
 }
 
 } // namespace utils
