+++ hawk-cli/src/helpers/utils.hpp
@@ -1,6 +1,7 @@
 #ifndef HAWK_CLI_UTILS_HPP
 #define HAWK_CLI_UTILS_HPP
 
+#include <cstddef>
 #include <cstdint>
 #include <optional>
 #include <string>
@@ -14,7 +15,7 @@ std::vector<std::string> tokenize(std::string_view str, bool quoted = true);
 
 std::optional<char> parse_delimiter(const std::string& str);
 
-std::uint8_t num_digits(std::uint64_t n);
+std::size_t num_digits(std::uint64_t n);
 
 } // namespace utils
 } // namespace hawk::cli
